feat: add working custom rsc support for delivery mode
