+++ packages/components/src/components/Assembly/Assembly.tsx
@@ -1,3 +1,4 @@
+'use client';
 import {
   ExperienceTreeNode,
   ResolveDesignValueType,
