+++ packages/components/src/components/Columns/Columns.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React, { CSSProperties, forwardRef } from 'react';
 import './Columns.css';
 import { combineClasses } from '../../utils/combineClasses';
