+++ packages/components/src/components/Text/Text.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { combineClasses } from '@/utils/combineClasses';
 import React from 'react';
 import './Text.css';
