+++ packages/core/src/types.ts
@@ -109,6 +109,7 @@ export type ComponentRegistration = {
   component: React.ElementType;
   definition: ComponentDefinition;
   options?: {
+    isRSC?: boolean;
     wrapComponent?: boolean;
     /** @deprecated use wrapContainer instead */
     wrapContainerTag?: keyof JSX.IntrinsicElements;
