+++ packages/core/src/utils/index.ts
@@ -12,3 +12,4 @@ export * from './pathSchema';
 export * from './resolveHyperlinkPattern';
 export * from './patternUtils';
 export * from './sanitizeNodeProps';
+export * from './getNodeProps';
