+++ packages/experience-builder-sdk/src/blocks/preview/PreviewUnboundImage.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React from 'react';
 import type {
   ComponentTreeNode,
