+++ packages/experience-builder-sdk/src/components/Flex.tsx
@@ -1,3 +1,4 @@
+'use client';
 import React, { MouseEventHandler, forwardRef } from 'react';
 import type * as CSS from 'csstype';
