+++ packages/test-apps/nextjs/src/studio-config.ts
@@ -29,6 +29,7 @@ defineComponents([
       },
     },
     options: {
+      isRSC: true,
       wrapComponent: false,
     },
   },
